+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Option `font.builtin_box_drawing` to disable the built-in font for drawing box characters
 - Track and report surface damage information to Wayland compositors
+- Builtin font is now used for characters from `U+e0b0` to `U+e0b3` usually correspond to powerline symbols
 
 ### Changed